requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
Django==5.1.5
filelock==3.17.0
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
instagrapi==2.1.5
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
MouseInfo==0.1.3
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
playwright-stealth==1.0.6
pluggy==1.5.0
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pyee==12.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.25.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
requests==2.32.3
selenium==4.28.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
tiktok-captcha-solver==0.6.2
trio==0.28.0
trio-websocket==0.11.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
urllib3==2.3.0
virtualenv==20.29.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wikipedia==1.4.0
wsproto==1.2.0
